+++ domain/src/main/java/com/twcc/domain/usecase/GetUsersUseCase.kt
@@ -1,11 +1,11 @@
 package com.twcc.domain.usecase
 
 import com.twcc.domain.models.UserDomain
-import com.twcc.domain.repository.Repository
+import com.twcc.domain.repository.UserRepository
 
-class GetUsersUseCase(private val repository: Repository) {
+class GetUsersUseCase(private val userRepository: UserRepository) {
 
     suspend fun execute(): List<UserDomain> {
-        return repository.getUsers()
+        return userRepository.getUsers()
     }
 }
